+++ lib/src/widgets/pspdfkit_flutter_widget_controller_impl.dart
@@ -1,3 +1,12 @@
+///
+///  Copyright © 2024-2025 PSPDFKit GmbH. All rights reserved.
+///
+///  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
+///  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
+///  UNAUTHORIZED REPRODUCTION OR DISTRIBUTION IS SUBJECT TO CIVIL AND CRIMINAL PENALTIES.
+///  This notice may not be removed from this file.
+///
+
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
@@ -298,5 +307,20 @@ class PspdfkitFlutterWidgetControllerImpl
   @override
   void onCustomToolbarItemTapped(String identifier) {
     onCustomToolbarItemTappedListener?.call(identifier);
+  }
+
+  @override
+  void addWebEventListener(
+      NutrientWebEvent event, Function(dynamic p1) callback) {
+    // Web event listeners are not supported on native platforms.
+    throw UnimplementedError('addWebEventListener is only supported on web.');
+  }
+
+  @override
+  void removeWebEventListener(
+      NutrientWebEvent event, Function(dynamic p1) callback) {
+    // Web event listeners are not supported on native platforms.
+    throw UnimplementedError(
+        'removeWebEventListener is only supported on web.');
   }
 }